requirements.txt
Flask==2.3.3
py-healthcheck==1.10.1


azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
python-dotenv==1.0.0